requirements.txt
django == 3.2
django-crispy-forms == 1.8.1
gunicorn == 19.9.0
whitenoise == 4.1.4
django-taggit == 1.2.0
django-tinymce
pillow == 2
django-utils-six
django-verify-email
django-user-accounts
djangorestframework == 3.11.0
requests
django-cors-headers